requirements.txt
fastapi
uvicorn
pymupdf
unstructured
sentence-transformers
chromadb
networkx
llama-index
trl
peft
ragas
gradio
streamlit
arxiv
PyYAML
scikit-learn
numpy
python-dotenv
requests
